Re #1: Initial cluster manager support
This commit creates a cluster manager which reads the mongodb information
created by the controller and generates kubernetes configs on the fly and
configures the cluster dynamically.

This commit also creates a comprehensive mechanism to get an entire nextensio
setup on our laptop - including the controller and ui/ux and the cluster with
minions and clustermgr - the whole nine yards. Refer to the readme in test/kind
for more details.

Testing/Verfication of clustermgr: the testing/verification is done using
the shorty.py test client which uses a two cluster setup. Very soon I will also
add support for the full agent/connector and browser/web-browsing via the
test cluster