Ignore the host name in the ingress gateway yaml
We will have case where agent tries to contact gateway.nextensio.net whereas
underneat it actually is connecting to gatewayuseast1.nextensio.net. So we dont
want the ingress gateway to enforce a particular host name. We can rather just
ask for it to look for gateway*.nextensio.net rather than "*", but thats for
another day