Support deletion and cleanup of every singe yaml whenever customer deletes something
This is enabled by the use of a "summary" collection that stores enough information in mongo
that indicates what was configured in k8s, that information is enough to rebuild yamls when
an element is deleted.

Also VERY CAREFULLY NOTE that the sequence of operations is critical. When adding a new config

1. First delete anything that might be not-needed now because of the new config
2. Now update the summary DB with the new config
3. Now update the new config

Also ensure that if we attempt to delete something that doesnt exist (NotFound), do not
complain and move on gracefully

There are enough comments in the code wherever the sequence has to be carefull observed.
There is a LOT of thought put in here to ensure that if clustermgr  crashes in the middle
of any operation, it should come up and complete the operations succesfully - whether it
be deleting removed stuff or adding new stuff. So when changing things in this area, make
sure to put a LOT of thought about all kinds of corner cases. We dont want a k8s cluster
with broken yamls !!